Add ContactType Patch method. Return 404 if resource is not found on Get and Delete.

+++ Part 3 Web API and OData V4 CRUD and Actions/WebAPIODataV4/Controllers/ContactTypeController.cs
@@ -24,7 +24,12 @@ namespace WebAPIODataV4.Controllers
         [EnableQuery(PageSize = 20)]
         public IHttpActionResult Get([FromODataUri] int key)
         {
-            return Ok(_db.ContactType.Find(key));
+            var contactType = _db.ContactType.Find(key);
+            if (contactType == null)
+            {
+                return NotFound();
+            }
+            return Ok(contactType);
         }
 
         [ODataRoute()]
@@ -66,11 +71,37 @@ namespace WebAPIODataV4.Controllers
         [ODataRoute]
         [HttpDelete]
         [EnableQuery]
-        public void Delete([FromODataUri] int key)
+        public IHttpActionResult Delete([FromODataUri] int key)
         {
             var entityInDb = _db.ContactType.SingleOrDefault(t => t.ContactTypeID == key);
+            if (entityInDb == null)
+            {
+                return NotFound();
+            }
             _db.ContactType.Remove(entityInDb);
             _db.SaveChanges();
+            return StatusCode(HttpStatusCode.NoContent);
+        }
+
+        [ODataRoute]
+        [HttpPatch]
+        [EnableQuery]
+        public IHttpActionResult Patch([FromODataUri] int key, Delta<ContactType> delta)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var contactType = _db.ContactType.SingleOrDefault(t => t.ContactTypeID == key);
+            if (contactType == null)
+            {
+                return NotFound();
+            }
+
+            delta.Patch(contactType);
+            _db.SaveChanges();
+            return Updated(contactType);
         }
 
         protected override void Dispose(bool disposing)
